Reapply "refactor: split local_fallback.py (1605 lines) into 7 sub-modules"
This reverts commit b1e4568a2caef252576e6135ad61f9ef8845181c.
