+++ client/.vite/deps/_metadata.json
@@ -1,0 +1,8 @@
+{
+  "hash": "a681efbd",
+  "configHash": "38a0a0fb",
+  "lockfileHash": "da55d71b",
+  "browserHash": "1fbef3b3",
+  "optimized": {},
+  "chunks": {}
+}
